Rolling back unnecessary code from chery pick prior to merging
