+++ qudi/logic/pulsed/sequence_generator_logic.py
@@ -32,7 +32,7 @@ from PySide2 import QtCore
 from qudi.core.statusvariable import StatusVar
 from qudi.core.connector import Connector
 from qudi.core.configoption import ConfigOption
-from qudi.core.paths import get_main_dir, get_home_dir
+from qudi.util.paths import get_main_dir, get_home_dir
 from qudi.util.helpers import natural_sort
 from qudi.util.network import netobtain
 from qudi.core.module import LogicBase
